Extracted attributes from "OfferItem" to new class "Discount"
